+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added `:authorization_params` config option to `PowAssent.Strategy.OAuth`
 * Plug and Phoenix controller now handles `:session_params` rather than `:state` for any params that needs to be stored temporarily during authorization
 * Added handling of `oauth_token_secret` to OAuth strategies
+* Support any `:plug` version below `2.0.0`
 
 ## v0.2.2 (2019-03-25)