+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Plug and Phoenix controller now handles `:session_params` rather than `:state` for any params that needs to be stored temporarily during authorization
 * Added handling of `oauth_token_secret` to OAuth strategies
 * Support any `:plug` version below `2.0.0`
+* Fixed bug in `mix pow_assent.ecto.gen.migration` task where `--binary-id` flag didn't generate correct migration
+* Support `:pow` version `1.0.5`
 
 ## v0.2.2 (2019-03-25)